refactor(ItemCount, ItemsTest): ItemCount 객체 생성 및 기존 결과의 수 검증 테스트 위치 변경
