+++ src/test/java/model/items/ItemsTest.java
@@ -20,16 +20,6 @@ class ItemsTest {
     }
 
     @Test
-    @DisplayName("참여 인원 수와 결과 목록의 수가 동일해야 한다.")
-    void checkItemCountWithPersonCount() {
-        List<String> items = List.of("꽝", "5000", "꽝", "3000");
-        PersonCount personCount = new PersonCount(3);
-        assertThatThrownBy(() -> Items.of(items, personCount))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("참여 인원 수와 결과 목록의 수가 동일하지 않습니다.");
-    }
-
-    @Test
     @DisplayName("사다리 타기를 완료한 후, 해당 결과의 인덱스에 위치한 사람에게 주어진다.")
     void receiveItemByIndex() {
         List<String> items = List.of("꽝", "5000", "꽝", "3000");
